Various cleanup - constant usage when possible - inline constants that are used in one location - minor todo and other cleanup in base
Issue#: CODE-2426


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@23293 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/base/formula/DividingFormula.java
@@ -87,7 +87,7 @@ public class DividingFormula implements ReferenceFormula<Integer>
 		 * rounding, and rounding first results in a faster & more accurate
 		 * calculation.
 		 */
-		return numbers[0].intValue() / denominator;
+		return Integer.valueOf(numbers[0].intValue() / denominator);
 	}
 
 	/**
